Update.
1998-08-20 20:07  Ulrich Drepper  <drepper@cygnus.com>

	* pwd/getpw.c (__getpw): Use %lu for uid and gid parameter and
	cast them to unsigned long int.
	* pwd/putpwent.c (putpwent): Likewise.
	Patch by John Tobey <jtobey@banta-im.com>.

1998-08-20 17:14  H.J. Lu  <hjl@gnu.org>

	* libio/libioP.h (FILEBUF_LITERAL): Set the _offset field to
	_IO_pos_BAD.

1998-02-20 17:54  H.J. Lu  <hjl@gnu.org>

	* libio/fileops.c (_IO_file_seekoff): Don't adjust pointers if
	_IO_SYSSEEK fails.

	* sysdeps/unix/sysv/linux/Makefile [subdirs=misc] (sysdep_headers):
	Add sys/sendfile.h.
	* sysdeps/unix/sysv/linux/syscalls.list: Add sendfile.
	* sysdeps/unix/sysv/linux/sys/sendfile.h: New file.

	* sysdeps/unix/sysv/linux/sys/acct.h: Pretty print.

	* sysdeps/unix/sysv/linux/Versions [GLIBC_2.1]: Add umount2 and
	sendfile.

+++ pwd/getpw.c
@@ -50,8 +50,9 @@ __getpw (uid, buf)
   if (__getpwuid_r (uid, &resbuf, tmpbuf, buflen, &p) < 0)
     return -1;
 
-  if (sprintf (buf, "%s:%s:%u:%u:%s:%s:%s", p->pw_name, p->pw_passwd,
-	       p->pw_uid, p->pw_gid, p->pw_gecos, p->pw_dir, p->pw_shell) < 0)
+  if (sprintf (buf, "%s:%s:%lu:%lu:%s:%s:%s", p->pw_name, p->pw_passwd,
+	       (unsigned long int) p->pw_uid, (unsigned long int) p->pw_gid,
+	       p->pw_gecos, p->pw_dir, p->pw_shell) < 0)
     return -1;
 
   return 0;
